Added services to controllers
